Resolves issues with pagination

+++ warden/group/manager_sql.go
@@ -142,7 +142,7 @@ func (m *SQLManager) RemoveGroupMembers(group string, subjects []string) error {
 	return nil
 }
 
-func (m *SQLManager) FindGroupsByMember(subject string, limit, offset int64) ([]Group, error) {
+func (m *SQLManager) FindGroupsByMember(subject string, limit, offset int) ([]Group, error) {
 	var ids []string
 	if err := m.DB.Select(&ids, m.DB.Rebind("SELECT group_id from hydra_warden_group_member WHERE member = ? GROUP BY group_id ORDER BY group_id LIMIT ? OFFSET ?"), subject, limit, offset); err == sql.ErrNoRows {
 		return nil, errors.WithStack(pkg.ErrNotFound)
@@ -163,7 +163,7 @@ func (m *SQLManager) FindGroupsByMember(subject string, limit, offset int64) ([]
 	return groups, nil
 }
 
-func (m *SQLManager) ListGroups(limit, offset int64) ([]Group, error) {
+func (m *SQLManager) ListGroups(limit, offset int) ([]Group, error) {
 	var ids []string
 	if err := m.DB.Select(&ids, m.DB.Rebind("SELECT group_id from hydra_warden_group_member GROUP BY group_id ORDER BY group_id LIMIT ? OFFSET ?"), limit, offset); err == sql.ErrNoRows {
 		return nil, errors.WithStack(pkg.ErrNotFound)
